Move hook to different folder.
